java.lang.invoke.MethodHandle's methods have magic return type Set the type of the expression be the expected type, so it isn't highlighted
See https://github.com/scala/scala/commit/c1d092a2fa2ffe08e7eb895501193aac6c2de5a2
 #SCL-9029 Fixed
