+++ test/org/jetbrains/plugins/scala/failed/annotator/JavaHighlightingTest.scala
@@ -73,34 +73,6 @@ class JavaHighlightingTest extends JavaHighlitghtingTestBase {
 
     assertNothing(errorsFromScalaCode(scala, java))
   }
-
-  def testSCL9029() = {
-    val scala =
-      """
-        |package scl9029
-        |import java.lang.invoke.{MethodHandles, MethodType}
-        |
-        |class SCL9029 {
-        |  def a: Int = 5
-        |
-        |  def b = {
-        |    val mh = MethodHandles.publicLookup().findVirtual(
-        |      classOf[A], "a", MethodType.methodType(classOf[Int])
-        |    )
-        |    val z: Int = mh.invokeExact(this)
-        |  }
-        |}
-      """.stripMargin
-
-    val java =
-      """
-        |package scl9029;
-        |public class Foo {
-        |}
-      """.stripMargin
-
-    assertNothing(errorsFromScalaCode(scala, java))
-  }
   
   def testSCL6409() = {
     val java =
